Make footnotes section an HTML comment
I noticed that the footnotes section appears as a visible heading when rendered as HTML (e.g. in GitHub), which looks odd because there's nothing below it, so I thought it might be useful to wrap it in an HTML comment so it's only seen by those who view it in its text form.  Feel free to close this PR if you disagree though!