delete Dockerfile: broken, and I think it's for the server anyways
The servers were moved out to separate repos (magic-wormhole-mailbox-server
and magic-wormhole-transit-relay), and I think this Dockerfile was building
an image mostly to run the mailbox server.

It might make sense in the future to have a Docker image for just the client,
in which case it'll be time to bring these Dockerfiles back with different
contents.

refs #295