Implement streaming zipfile creation for sending directories
Instead of creating a temporary zip file on disk to send when
transferring a directory, this commit uses the zipstream-ng library to
generate the zip file on the fly. This results in no disk usage, a much
shorter startup time, and minimal memory usage.

For adding files to the stream from the directory, the `walk` function
the `zipstream-ng` package provides was used since it preserves empty
directories as well as follows symlinks.

Since files aren't actually read until they're about to be transferred,
an `os.access` check is done to make sure that unreadable files are not
added to the stream (which would cause it to crash mid-transfer).