Remove expectation of zip files using deflate compression
The check that directories are expected to have a mode of
"zipfile/deflate" was changed to only require that the mode starts with
"zipfile". The tests were also updated.

In order to preserve backwards compatibility, the mode used when sending
a directory was left as the now-incorrect "zipfile/deflate"

Since the compression method passed to the ZipFile class is only used
when writing files, it was removed in all cases where the zip is being
opened in read-only mode.